requirements.txt
selectolax
aiohttp
tenacity
lxml
PyYAML
psycopg2-binary
requests
google-genai
